backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic[email]==2.7.1
anthropic==0.28.0
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
openpyxl==3.1.2
